+++ R/tar_load_archive.R
@@ -1,0 +1,44 @@
+#' Load a target's value from archive storage
+#'
+#' tarchives version of [targets::tar_load()]. Reads one or more targets from
+#' an archived pipeline's store and assigns them into an environment.
+#'
+#' @param package A scalar character of the package name.
+#' @param pipeline A scalar character of the pipeline name.
+#' @inheritParams targets::tar_load
+#'
+#' @inherit targets::tar_load return
+#'
+#' @export
+tar_load_archive <- function(
+  names,
+  package,
+  pipeline,
+  branches = NULL,
+  meta = NULL,
+  strict = TRUE,
+  silent = FALSE,
+  envir = parent.frame(),
+  store = targets::tar_config_get("store")
+) {
+  check_string(package, allow_empty = FALSE)
+  check_string(pipeline, allow_empty = FALSE)
+  store <- tar_archive_store(
+    package = package,
+    pipeline = pipeline,
+    store = store
+  )
+  meta <- meta %||%
+    targets::tar_meta(
+      store = store
+    )
+  targets::tar_load(
+    names = {{ names }},
+    branches = branches,
+    meta = meta,
+    strict = strict,
+    silent = silent,
+    envir = envir,
+    store = store
+  )
+}