+++ tests/testthat/test-tar_meta_archive.R
@@ -1,0 +1,12 @@
+targets::tar_test("tar_meta_archive() returns metadata", {
+  tar_make_archive(
+    package = "tarchives",
+    pipeline = "example-model"
+  )
+
+  meta <- tar_meta_archive(
+    package = "tarchives",
+    pipeline = "example-model"
+  )
+  expect_in("model", meta$name)
+})